+++ mmpose/models/__init__.py
@@ -3,9 +3,8 @@ from .builder import (BACKBONES, HEADS, LOSSES, NECKS, POSENETS,
                       build_backbone, build_head, build_loss, build_neck,
                       build_posenet)
 from .detectors import *  # noqa
-from .keypoint_heads import *  # noqa
+from .heads import *  # noqa
 from .losses import *  # noqa
-from .mesh_heads import *  # noqa
 from .necks import *  # noqa
 
 __all__ = [
