+++ mmpose/models/detectors/mesh.py
@@ -1,9 +1,9 @@
 import numpy as np
 import torch
 
+from mmpose.models.heads.discriminator import SMPLDiscriminator
 from .. import builder
 from ..builder import POSENETS
-from ..mesh_heads.discriminator import SMPLDiscriminator
 from .base import BasePose
 
 try:
