+++ mmpose/models/heads/__init__.py
@@ -1,6 +1,7 @@
 from .ae_higher_resolution_head import AEHigherResolutionHead
 from .ae_simple_head import AESimpleHead
 from .deeppose_regression_head import DeepposeRegressionHead
+from .hmr_head import HMRMeshHead
 from .interhand_3d_head import Interhand3DHead
 from .temporal_regression_head import TemporalRegressionHead
 from .topdown_heatmap_base_head import TopdownHeatmapBaseHead
@@ -12,5 +13,5 @@ __all__ = [
     'TopdownHeatmapSimpleHead', 'TopdownHeatmapMultiStageHead',
     'TopdownHeatmapMSMUHead', 'TopdownHeatmapBaseHead',
     'AEHigherResolutionHead', 'AESimpleHead', 'DeepposeRegressionHead',
-    'TemporalRegressionHead', 'Interhand3DHead'
+    'TemporalRegressionHead', 'Interhand3DHead', 'HMRMeshHead'
 ]
