+++ mmpose/models/heads/topdown_heatmap_multi_stage_head.py
