+++ mmpose/models/heads/topdown_heatmap_simple_head.py
