--- mmpose/models/mesh_heads/__init__.py
@@ -1,3 +1,0 @@
-from .hmr_head import MeshHMRHead
-
-__all__ = ['MeshHMRHead']
